Fix token read from file
When a token is read from a file, the end-of-line char is also read,
unless the file is explicitly saved without it. What happens is that
the end-of-line will break the curl query alter one and cause a quite
cryptic message about the token not being correct.

This change makes sure any end-of-line or carriage-return character are removed
from the token.

Fixes #18