feat(config): accept to run llm-ls in PATH (#116)
* feat(config): accept to run llm-ls in PATH

It's uncommon to completely ignore PATH. I have llm-ls installed via my package manager, and the downloaded one wouldn't work anyway. If bin_path is not set, llm.nvim will try to start llm-ls from PATH, else fallback to config.lsp.bin_path. Note that we dont check The 'version' of llm-ls so it's possible for the plugin to run a version that doesn't match the one configured (for instance if you've added it after the first run of llm.nvim). A version check could/should be added later.

---------

Co-authored-by: Luc Georges <McPatate@users.noreply.github.com>